Renamed version.py to autover.py
